fix(generator): Avoid compile-time filter map querying when generating static parser.
This fixes https://github.com/angular/angular.dart/issues/432.

Also consistenly load libraries through the package root to avoid getting in
trouble if the same library is loaded through different URIs.
