+++ lib/playback/playback_http.dart
@@ -8,7 +8,7 @@ import 'package:angular/core_dom/module.dart';
 import 'package:angular/core/service.dart';
 import 'package:angular/mock/module.dart' as mock;
 
-import 'playback_data.dart' as playback_data;
+import 'package:angular/playback/playback_data.dart' as playback_data;
 
 @NgInjectableService()
 class PlaybackHttpBackendConfig {
