+++ lib/tools/expression_extractor.dart
@@ -10,8 +10,10 @@ import 'package:angular/tools/common.dart';
 
 import 'package:di/di.dart';
 import 'package:di/dynamic_injector.dart';
-import 'package:angular/tools/parser_generator/generator.dart';
+
+import 'package:angular/core/module.dart';
 import 'package:angular/core/parser/parser.dart';
+import 'package:angular/tools/parser_generator/generator.dart';
 
 main(args) {
   if (args.length < 5) {
@@ -50,9 +52,10 @@ main(args) {
 
   printer.printSrc('// Found ${expressions.length} expressions');
   Module module = new Module()
-    ..type(Parser, implementedBy: DynamicParser)
-    ..type(ParserBackend, implementedBy: DynamicParserBackend)
-    ..value(SourcePrinter, printer);
+      ..type(Parser, implementedBy: DynamicParser)
+      ..type(ParserBackend, implementedBy: DynamicParserBackend)
+      ..type(FilterMap, implementedBy: NullFilterMap)
+      ..value(SourcePrinter, printer);
   Injector injector =
       new DynamicInjector(modules: [module], allowImplicitInjection: true);
