+++ lib/tools/source_metadata_extractor.dart
@@ -2,9 +2,9 @@ library angular.source_metadata_extractor ;
 
 import 'package:analyzer/src/generated/ast.dart';
 
-import 'source_crawler.dart';
-import '../utils.dart';
-import 'common.dart';
+import 'package:angular/tools/source_crawler.dart';
+import 'package:angular/tools/common.dart';
+import 'package:angular/utils.dart';
 
 const String _COMPONENT = '-component';
 const String _DIRECTIVE = '-directive';
